RF: core AnnexRepo class
And a whole slew of changes that was necessary to be able to run
AnnexRepo.is_valid_annex() in this new class.
