+++ datalad/customremotes/tests/test_archives.py
@@ -144,14 +144,14 @@ def test_basic_scenario(d, d2):
 
     # Check if protocol was collected
     if os.environ.get('DATALAD_TESTS_PROTOCOLREMOTE'):
-        assert_is_instance(annex.cmd_call_wrapper.protocol, AnnexExchangeProtocol)
+        assert_is_instance(annex._cmd_call_wrapper.protocol, AnnexExchangeProtocol)
         protocol_file = _path_(annex.path,
                                '.git/bin/git-annex-remote-datalad-archive')
         ok_file_has_content(protocol_file, "VERSION 1", re_=True, match=False)
         ok_file_has_content(protocol_file, "GETAVAILABILITY", re_=True, match=False)
         ok_file_has_content(protocol_file, "#!/bin/bash", re_=True, match=False)
     else:
-        assert_false(isinstance(annex.cmd_call_wrapper.protocol, AnnexExchangeProtocol))
+        assert_false(isinstance(annex._cmd_call_wrapper.protocol, AnnexExchangeProtocol))
 
     # verify that we can drop if original archive gets dropped but available online:
     #  -- done as part of the test_add_archive_content.py
